+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ lipyphilic CHANGELOG
 0.1.0 (????-??-??)
 ------------------
 
+* Add functionality to find flip-flop events in bilayers
 * Add functionality to calculate area per lipid
 * Add functionality to find assign lipids to leaflets in a bilayer